refactor: helpers
