+++ index.js
@@ -1,3 +1,4 @@
+const _ = require('lodash');
 const { getModelSegments } = require('./lib');
 
 /**
