+++ lib/index.js
@@ -1,127 +1,8 @@
 const _ = require('lodash');
 const m2s = require('mongoose-to-swagger');
-
-const queryParameters = [
-  {
-    name: 'sort',
-    in: 'query',
-    schema: {
-      type: 'string',
-    },
-  },
-  {
-    name: 'skip',
-    in: 'query',
-    schema: {
-      type: 'number',
-    },
-  },
-  {
-    name: 'limit',
-    in: 'query',
-    schema: {
-      type: 'number',
-    },
-  },
-  {
-    name: 'query',
-    in: 'query',
-    schema: {
-      type: 'string',
-    },
-  },
-  {
-    name: 'populate',
-    in: 'query',
-    schema: {
-      type: 'string',
-    },
-  },
-  {
-    name: 'select',
-    in: 'query',
-    schema: {
-      type: 'string',
-    },
-  },
-  {
-    name: 'distinct',
-    in: 'query',
-    schema: {
-      type: 'string',
-    },
-  },
-];
-
-const pathParamaters = [
-  {
-    name: 'id',
-    in: 'path',
-    required: true,
-    schema: {
-      type: 'string',
-    },
-  },
-];
-
-const countSchema = {
-  type: 'object',
-  properties: {
-    count: {
-      type: 'number',
-    },
-  },
-  required: ['count'],
-};
-
-const errors = {
-  default: {
-    description: 'Unexpected Error',
-    content: {
-      'application/json': {
-        schema: {
-          $ref: '#/components/schemas/error',
-        },
-      },
-    },
-  },
-  badRequest: {
-    description: 'Bad Request',
-    content: {
-      'application/json': {
-        schema: {
-          $ref: '#/components/schemas/error',
-        },
-      },
-    },
-  },
-  notFound: {
-    description: 'Not Found',
-    content: {
-      'application/json': {
-        schema: {
-          $ref: '#/components/schemas/error',
-        },
-      },
-    },
-  },
-};
-
-const getOperationId = ({ method = 'get', model, id = false, subcommand }) => {
-  const operation = {
-    get: 'get',
-    post: id ? 'update' : 'create',
-    put: 'replace',
-    patch: 'patch',
-    delete: 'delete',
-  };
-
-  const prefix = operation[method.toLowerCase()];
-  const suffix = method === 'get' && id ? 'byId' : '';
-  const parts = [prefix, _.startCase(model), suffix, subcommand];
-
-  return _.chain(parts).compact().join(' ').camelCase().value();
-};
+const { getErrorResponses } = require('./errors');
+const { getQueryParameters, getPathParameters } = require('./parameters');
+const { getOperationId } = require('./operations');
 
 const getModelSchemas = ([
   key,
@@ -156,7 +37,7 @@ const getModelPaths = (
         summary: `Query ${key}`,
         operationId: getOperationId({ model: key }),
         tags: [key],
-        parameters: [...queryParameters],
+        parameters: getQueryParameters(),
         responses: {
           200: {
             description: 'OK',
@@ -171,12 +52,7 @@ const getModelPaths = (
               },
             },
           },
-          400: {
-            ...errors.badRequest,
-          },
-          default: {
-            ...errors.default,
-          },
+          ...getErrorResponses([400]),
         },
         ...globalPathOverrides,
         ...modelPathOverrides,
@@ -206,12 +82,7 @@ const getModelPaths = (
               },
             },
           },
-          400: {
-            ...errors.badRequest,
-          },
-          default: {
-            ...errors.default,
-          },
+          ...getErrorResponses([400]),
         },
         ...globalPathOverrides,
         ...modelPathOverrides,
@@ -225,24 +96,25 @@ const getModelPaths = (
           subcommand: 'count',
         }),
         tags: [key],
-        parameters: [...queryParameters],
+        parameters: getQueryParameters(),
         responses: {
           200: {
             description: 'OK',
             content: {
               'application/json': {
                 schema: {
-                  ...countSchema,
+                  type: 'object',
+                  properties: {
+                    count: {
+                      type: 'number',
+                    },
+                  },
+                  required: ['count'],
                 },
               },
             },
           },
-          400: {
-            ...errors.badRequest,
-          },
-          default: {
-            ...errors.default,
-          },
+          ...getErrorResponses([400]),
         },
         ...globalPathOverrides,
         ...modelPathOverrides,
@@ -253,7 +125,7 @@ const getModelPaths = (
         summary: `Get ${key} by id`,
         operationId: getOperationId({ model: key, id: true }),
         tags: [key],
-        parameters: [...pathParamaters],
+        parameters: getPathParameters(),
         responses: {
           200: {
             description: 'OK',
@@ -265,12 +137,7 @@ const getModelPaths = (
               },
             },
           },
-          404: {
-            ...errors.notFound,
-          },
-          default: {
-            ...errors.default,
-          },
+          ...getErrorResponses([400, 404]),
         },
         ...globalPathOverrides,
         ...modelPathOverrides,
@@ -283,7 +150,7 @@ const getModelPaths = (
           model: key,
         }),
         tags: [key],
-        parameters: [...pathParamaters],
+        parameters: getPathParameters(),
         requestBody: {
           content: {
             'application/json': {
@@ -305,15 +172,7 @@ const getModelPaths = (
               },
             },
           },
-          400: {
-            ...errors.badRequest,
-          },
-          404: {
-            ...errors.notFound,
-          },
-          default: {
-            ...errors.default,
-          },
+          ...getErrorResponses([400, 404]),
         },
         ...globalPathOverrides,
         ...modelPathOverrides,
@@ -326,7 +185,7 @@ const getModelPaths = (
           model: key,
         }),
         tags: [key],
-        parameters: [...pathParamaters],
+        parameters: getPathParameters(),
         requestBody: {
           content: {
             'application/json': {
@@ -348,15 +207,7 @@ const getModelPaths = (
               },
             },
           },
-          400: {
-            ...errors.badRequest,
-          },
-          404: {
-            ...errors.notFound,
-          },
-          default: {
-            ...errors.default,
-          },
+          ...getErrorResponses([400, 404]),
         },
         ...globalPathOverrides,
         ...modelPathOverrides,
@@ -369,7 +220,7 @@ const getModelPaths = (
           model: key,
         }),
         tags: [key],
-        parameters: [...pathParamaters],
+        parameters: getPathParameters(),
         requestBody: {
           content: {
             'application/json': {
@@ -391,15 +242,7 @@ const getModelPaths = (
               },
             },
           },
-          400: {
-            ...errors.badRequest,
-          },
-          404: {
-            ...errors.notFound,
-          },
-          default: {
-            ...errors.default,
-          },
+          ...getErrorResponses([400, 404]),
         },
         ...globalPathOverrides,
         ...modelPathOverrides,
@@ -412,17 +255,12 @@ const getModelPaths = (
           id: true,
         }),
         tags: [key],
-        parameters: [...pathParamaters],
+        parameters: getPathParameters(),
         responses: {
           204: {
             description: 'Deleted',
           },
-          404: {
-            ...errors.notFound,
-          },
-          default: {
-            ...errors.default,
-          },
+          ...getErrorResponses([404]),
         },
         ...globalPathOverrides,
         ...modelPathOverrides,
@@ -437,7 +275,7 @@ const getModelPaths = (
           subcommand: 'shallow',
         }),
         tags: [key],
-        parameters: [...pathParamaters],
+        parameters: getPathParameters(),
         responses: {
           200: {
             description: 'OK',
@@ -449,12 +287,7 @@ const getModelPaths = (
               },
             },
           },
-          404: {
-            ...errors.notFound,
-          },
-          default: {
-            ...errors.default,
-          },
+          ...getErrorResponses([404]),
         },
         ...globalPathOverrides,
         ...modelPathOverrides,
