+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "erm2openapi",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "Create OpenAPI documentation from Mongoose schemas for express-restify-mongoose servers.",
   "main": "index.js",
   "scripts": {
@@ -34,8 +34,11 @@
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jest": "^24.3.2",
     "eslint-plugin-prettier": "^3.3.1",
+    "graphql": "^15.5.0",
     "jest": "^26.6.3",
     "mongoose": "^5.12.0",
+    "oas-validator": "^5.0.5",
+    "openapi-to-graphql": "^2.3.0",
     "prettier": "^2.2.1"
   },
   "dependencies": {
