+++ test/fixtures/api-docs.json
@@ -15,9 +15,7 @@
       "get": {
         "summary": "Query story",
         "operationId": "getStory",
-        "tags": [
-          "story"
-        ],
+        "tags": ["story"],
         "parameters": [
           {
             "name": "sort",
@@ -108,9 +106,7 @@
       "post": {
         "summary": "Create new story",
         "operationId": "createStory",
-        "tags": [
-          "story"
-        ],
+        "tags": ["story"],
         "requestBody": {
           "content": {
             "application/json": {
@@ -159,9 +155,7 @@
       "get": {
         "summary": "Get story count",
         "operationId": "getStoryCount",
-        "tags": [
-          "story"
-        ],
+        "tags": ["story"],
         "parameters": [
           {
             "name": "sort",
@@ -225,9 +219,7 @@
                       "type": "number"
                     }
                   },
-                  "required": [
-                    "count"
-                  ]
+                  "required": ["count"]
                 }
               }
             }
@@ -259,9 +251,7 @@
       "get": {
         "summary": "Get story by id",
         "operationId": "getStoryById",
-        "tags": [
-          "story"
-        ],
+        "tags": ["story"],
         "parameters": [
           {
             "name": "id",
@@ -279,6 +269,16 @@
               "application/json": {
                 "schema": {
                   "$ref": "#/components/schemas/story"
+                }
+              }
+            }
+          },
+          "400": {
+            "description": "Bad Request",
+            "content": {
+              "application/json": {
+                "schema": {
+                  "$ref": "#/components/schemas/error"
                 }
               }
             }
@@ -308,9 +308,7 @@
       "post": {
         "summary": "Update story by id",
         "operationId": "updateStory",
-        "tags": [
-          "story"
-        ],
+        "tags": ["story"],
         "parameters": [
           {
             "name": "id",
@@ -377,9 +375,7 @@
       "put": {
         "summary": "Update story by id",
         "operationId": "replaceStory",
-        "tags": [
-          "story"
-        ],
+        "tags": ["story"],
         "parameters": [
           {
             "name": "id",
@@ -446,9 +442,7 @@
       "patch": {
         "summary": "Update story by id",
         "operationId": "patchStory",
-        "tags": [
-          "story"
-        ],
+        "tags": ["story"],
         "parameters": [
           {
             "name": "id",
@@ -515,9 +509,7 @@
       "delete": {
         "summary": "Delete story by id",
         "operationId": "deleteStory",
-        "tags": [
-          "story"
-        ],
+        "tags": ["story"],
         "parameters": [
           {
             "name": "id",
@@ -559,9 +551,7 @@
       "get": {
         "summary": "Get story by id shallow",
         "operationId": "getStoryByIdShallow",
-        "tags": [
-          "story"
-        ],
+        "tags": ["story"],
         "parameters": [
           {
             "name": "id",
@@ -610,9 +600,7 @@
       "get": {
         "summary": "Query person",
         "operationId": "getPerson",
-        "tags": [
-          "person"
-        ],
+        "tags": ["person"],
         "parameters": [
           {
             "name": "sort",
@@ -703,9 +691,7 @@
       "post": {
         "summary": "Create new person",
         "operationId": "createPerson",
-        "tags": [
-          "person"
-        ],
+        "tags": ["person"],
         "requestBody": {
           "content": {
             "application/json": {
@@ -754,9 +740,7 @@
       "get": {
         "summary": "Get person count",
         "operationId": "getPersonCount",
-        "tags": [
-          "person"
-        ],
+        "tags": ["person"],
         "parameters": [
           {
             "name": "sort",
@@ -820,9 +804,7 @@
                       "type": "number"
                     }
                   },
-                  "required": [
-                    "count"
-                  ]
+                  "required": ["count"]
                 }
               }
             }
@@ -854,9 +836,7 @@
       "get": {
         "summary": "Get person by id",
         "operationId": "getPersonById",
-        "tags": [
-          "person"
-        ],
+        "tags": ["person"],
         "parameters": [
           {
             "name": "id",
@@ -874,6 +854,16 @@
               "application/json": {
                 "schema": {
                   "$ref": "#/components/schemas/person"
+                }
+              }
+            }
+          },
+          "400": {
+            "description": "Bad Request",
+            "content": {
+              "application/json": {
+                "schema": {
+                  "$ref": "#/components/schemas/error"
                 }
               }
             }
@@ -903,9 +893,7 @@
       "post": {
         "summary": "Update person by id",
         "operationId": "updatePerson",
-        "tags": [
-          "person"
-        ],
+        "tags": ["person"],
         "parameters": [
           {
             "name": "id",
@@ -972,9 +960,7 @@
       "put": {
         "summary": "Update person by id",
         "operationId": "replacePerson",
-        "tags": [
-          "person"
-        ],
+        "tags": ["person"],
         "parameters": [
           {
             "name": "id",
@@ -1041,9 +1027,7 @@
       "patch": {
         "summary": "Update person by id",
         "operationId": "patchPerson",
-        "tags": [
-          "person"
-        ],
+        "tags": ["person"],
         "parameters": [
           {
             "name": "id",
@@ -1110,9 +1094,7 @@
       "delete": {
         "summary": "Delete person by id",
         "operationId": "deletePerson",
-        "tags": [
-          "person"
-        ],
+        "tags": ["person"],
         "parameters": [
           {
             "name": "id",
@@ -1154,9 +1136,7 @@
       "get": {
         "summary": "Get person by id shallow",
         "operationId": "getPersonByIdShallow",
-        "tags": [
-          "person"
-        ],
+        "tags": ["person"],
         "parameters": [
           {
             "name": "id",
@@ -1205,10 +1185,7 @@
   "components": {
     "schemas": {
       "error": {
-        "required": [
-          "name",
-          "message"
-        ],
+        "required": ["name", "message"],
         "type": "object",
         "properties": {
           "name": {
@@ -1221,9 +1198,7 @@
       },
       "story": {
         "type": "object",
-        "required": [
-          "title"
-        ],
+        "required": ["title"],
         "properties": {
           "author": {
             "type": "string"
@@ -1244,9 +1219,7 @@
       },
       "person": {
         "type": "object",
-        "required": [
-          "name"
-        ],
+        "required": ["name"],
         "properties": {
           "name": {
             "type": "string"
